In a working state, ish.
seperated color variables into max, submax and badmax.

badmax means that this quadrant/face is just a bad combination and
plasma climbs are affected.

submax is almost as good as the optimal angle.

max is the optimal angle, and different to badmax it is the best we can
do in quake.

Still some gaps between the submax and max pitch angles.